Adds version check and command line args to override statically set aws login credentials

+++ aws_s3_logs.rb
@@ -1,27 +1,55 @@
 #!/usr/bin/env ruby
 
+# 
+# **author:** R. Nick McCandless
+#
+# Command line script to download logs from S3 bucket, concatenate them info a file, formats the file, and analyzes it
+#
+# **Usages**
+# `$ ruby aws_s3_logs.rb <bucket_name> <access_key_id> <secret_access_key> <prefix>`
+# or
+# Hard code the variables @ the top of the file, then `$ ruby aws_s3_logs.rb`
+#
+
+# Ruby version check
+abort("You're using ruby #{RUBY_VERSION}. Please use version 2.1 or newer") if (RUBY_VERSION.to_f < 2.1)
+
+bucket_name           = ''
+access_key_id         = ''
+secret_access_key     = ''
+prefix                = ''
+
 require 'rubygems'
 require 'fileutils'
 require 'csv'
 require 'json'
 
+# prefix: Subfolder - will create subfolder on local drive similar to prefix
 aws_information = {
-    prefix:             '', # Subfolder - will create subfolder on local drive similar to prefix
-    bucket_name:        '',
-    access_key_id:      '',
-    secret_access_key:  ''
+    bucket_name:       ARGV[0] || bucket_name,
+    access_key_id:     ARGV[1] || access_key_id,
+    secret_access_key: ARGV[2] || secret_access_key,
+    prefix:            ARGV[3] || prefix
 }
 options = {
     aws_logs_dir: 'aws_s3_logs',
     concat_file:  "#{aws_information[:bucket_name]}.log"
 }
 
-puts 'What would you like to do with the S3 logs? download, concat, format, analyze, or exit'
-input1 = gets.chomp
+if (aws_information.any? { |k, v| !v.empty? })
+  puts "\nThe following credentials have been added:"
+  aws_information.each { |k, v| puts "#{k}: #{v}" }
+else
+  puts 'Please check usages - either pass in the S3 credentials in the command line or hard code them in the file.'
+  abort "\nNo S3 credentials provided - existing!\n"
+end
+
+puts "\n\nWhat would you like to do with the S3 logs? download, concat (or concatenate), format, analyze, or exit"
+input1 = $stdin.gets.chomp
 
 begin
   case input1
-    when 'download', 'concat', 'format', 'analyze'
+    when 'download', 'concat', 'concatenate', 'format', 'analyze'
       send(input1, aws_information, options)
     when 'exit', 'quit'
       abort 'Good bye!'
@@ -44,7 +72,7 @@ BEGIN {
     begin
       require 'aws-sdk'
     rescue LoadError
-      abort 'Unable to load AWS gem. Is the gem install?'
+      abort 'Unable to load AWS gem. Is the gem `aws-sdk` installed?'
     end
 
     s3 = Aws::S3::Client.new(
@@ -66,7 +94,6 @@ BEGIN {
       File.open("./aws_s3_logs/#{o.key}.txt", 'w') do |f|
         f.write resp.body.read
       end
-      # puts resp.body.read
 
       puts "Downloaded log file: #{o.key}"
     end
@@ -76,7 +103,7 @@ BEGIN {
 
 
   # Concatenate all log files into one
-  def concat aws_information, options
+  def concatenate aws_information, options
     puts "Concatenating all log files to log file named: #{options[:concat_file]}"
 
     log_count = Dir[File.join("./#{options[:aws_logs_dir]}/#{aws_information[:prefix]}", '**', '*')].length
@@ -87,7 +114,7 @@ BEGIN {
     end
 
     #Clear concat file first
-    File.truncate("./#{options[:aws_logs_dir]}/#{options[:concat_file]}", 0)
+    File.truncate("./#{options[:aws_logs_dir]}/#{options[:concat_file]}", 0) if File.exist?("./#{options[:aws_logs_dir]}/#{options[:concat_file]}")
 
     # Concatenate log files into one
     Dir.foreach(File.join("./#{options[:aws_logs_dir]}/#{aws_information[:prefix]}")) do |log|
@@ -97,7 +124,10 @@ BEGIN {
       end
     end
 
+    puts "\nConcatenate complete.\n"
   end
+
+  (class << self; self; end).send :alias_method, :concat, :concatenate
 
 
   # Formats logs
@@ -123,16 +153,19 @@ BEGIN {
       log_results.each { |l| csv << l.values }
     end
 
+    puts "\nFormat complete.\n"
   end
 
 
-  # Creates an analysis report
+  # Creates an analysis report to JSON & outputs the results 
   def analyze  aws_information, options
     analytics_results = {}
 
     log_results = concat_to_array_hash "./#{options[:aws_logs_dir]}/#{options[:concat_file]}", [:ip_address]
 
     analytics_results[:ip_count] = log_results.each_with_object(Hash.new(0)) { |word,counts| counts[word] += 1 }.sort_by{ |k, v| v }.reverse.to_h.first(100).map{ |v| v }.map { |key, val| {key[:ip_address] => val} }
+
+    puts "\nResults give are the top 100 IP sorted by count"
 
     # To JSON file
     File.open("./#{options[:aws_logs_dir]}/#{options[:concat_file]}_analysis.json","w") do |f|
@@ -159,6 +192,7 @@ BEGIN {
       request_data:     /\s\".*$/
     }
 
+    # Reads concatenated file and stores it in an array of hashes
     File.readlines(concat_file).each do |l|
       log_match = Hash.new
 
@@ -173,11 +207,6 @@ BEGIN {
 
 }
 
-
-
-at_exit {
-  puts "\n"
-}
 
 
 # EOD
